+++ examples/image/referential_image_generation.py
@@ -5,14 +5,14 @@ dotenv.load_dotenv()
 
 img = ImagePollinations()
 data = img.generate(
-    "Convierte esta foto en estilo ghibli, manteniendo la pose",
+    "Convierte esta foto en estilo pixar, manteniendo la pose",
     params={
         "model": "klein",
         "image": "https://upload.wikimedia.org/wikipedia/commons/thumb/8/8e/Ovejera_Alem%C3%A1n.jpg/500px-Ovejera_Alem%C3%A1n.jpg",
-        "width": 1024,
-        "height": 1024,
+        "width": 512,
+        "height": 512,
     },
 )
 
-with open("anime.png", "wb") as f:
+with open("pixar.png", "wb") as f:
     f.write(data)